feat: 1등 로또 한개 결과 구하기 구현

+++ src/test/java/LottoMachineTest.java
@@ -1,5 +1,6 @@
 import static org.assertj.core.api.Assertions.assertThat;
 
+import java.math.BigInteger;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -13,7 +14,7 @@ public class LottoMachineTest {
 
         LottoResult lottoResult = lottoMachine.getResult(List.of(List.of(1,2,3,4,5,6)));
 
-        assertThat(lottoResult.getTotalPrizeAmount()).isEqualTo(LottoRank.FIRST.getPrizeAmount());
+        assertThat(lottoResult.getTotalPrizeAmount()).isEqualTo(BigInteger.valueOf(LottoRank.FIRST.getPrizeAmount()));
         assertThat(lottoResult.getCountByRank(LottoRank.FIRST)).isEqualTo(1);
     }
 }
